requirements.txt
hidapi
psutil
pystray
Pillow
pyinstaller
